move dynamic imports out of the backbone init flow
we've got the backbone page showing the loading spinner forever, and the
methods initialized inside `continueInit` are never getting fired, so I
think there are timing issues at play.

I'm not sure this will work, but moving the dynamic imports stuff to
later than the rest of the router init. I don't know how we can pause to
load those routes. TBD if this will work.